reorganise modularity of code
turns out the class declarations all need to be in the header file. TIL.
