+++ xo_AI.cpp
@@ -1,59 +1,3 @@
-//contains
-class Tree;//prototype declaration
-
-class Computer: public Player //computer is a subclass of player
-{
-    protected:
-        Game_Tree* tree;
-        char colour;
-        double ;//if randomness is non zero, then the computer will sometimes pick non optimal moves
-    public:
-        Computer(Game_state* p_game, char player_colour, bool playergoesfirst):  tree(p_game, playergoesfirst); //will construct the computer object to play as the opposite colour. Also call the game_tree constructor.
-        bool make_move (Game_State* p_game);//changes the object passed to it to reflect the new board state
-        
-};
-
-
-class Tree_Node
-{
-    protected:
-       Game_state* state, temp;
-  
-
-       //bool max_or_min; //true if this node is a maximum of the previous nodes.
-       double value; //+10 for victory from this node, -10 if this node causes loss.
-                            //eventually i plan to a wider variety of numbers here.
-      
-       int depth; //0 for top level nodes, all the way down to 8 for terminating nodes.
-
-       //At each ply, there are potentially up to 8 different ways the game can develop
-       Tree_Node leaves [8]; //array of pointers to the potentially up to 8 different game leafs that are below this game leaf.
-                                    //some of the pointers might be NULL if moves cannot be played there
-
-    public:
-        Tree_Node(Tree_Node* current_leaf,  game_state* p_game, bool maximize, bool pgoesfirst); //constructor. second argument is an array of pointers to the lower down nodes
-        
-    //Tree_Node(Tree_Node* current_leaf, game_state* p_game, int value); //constructor for a terminating node.
-
-        double get_node_value(void);
-        Tree_node* get_lower_node (int i);//returns leaves[i]
-        
-        //double get_next_node(int depth);//returns a rarandom leaf at 
-};
-class Tree//will store the best possible move at each node
-{
-        Tree_Node* head;//pointer to the first leaf in the game tree
-        Tree_Node* current_leaf;//iterator
-        bool playergoesfirst;//true if player is X, false if computer is X
-        //int minimax (bool maximising);//used in constructing the game tree.
-   public:
-        Tree(Game_State* p_game, bool pgoesfirst); //constructor. 
-        
-        double get_node(int depth, int i);//returns the next node at a given depth.
-  
-        //Game_Leaf* picknextmove(double randomness);//returns the next play the computer should make
-                                                    //if randomness is non zero, then the computer will sometimes pick non optimal moves
-};
 
 
 
@@ -67,30 +11,30 @@ Tree::Tree( Game_state* p_game, bool pgoesfirst)//constructor
 }
 
 Tree_Node::Tree_Node(Tree_Node* current_leaf,  game_state* p_game, bool maximize, bool pgoesfirst)
-//constructor. 
+//constructor.
  {
-    
+
     depth= p_game->depth;
-    
+
     //create some variables that will be used in calcualting the node value
     int win_value=1;//this is the wieght that will be assigned for a win
     double best_value=0, test_value;
-    
+
     int index=1;
-    
-    //we can work out what colour the curerent player is by considering the fact that X alwatys plays first, 
+
+    //we can work out what colour the curerent player is by considering the fact that X alwatys plays first,
     //and the computer will always be trying to maximize
     char colour;
     if (maximize== pgoesfirst)
         colour= 'O';
     else
         colour= 'X';
-    
+
     //copy the input game state into this class.
-    //note that i have written a cutom copy operator in xo_functions.cpp 
-    *state = *p_game; 
+    //note that i have written a cutom copy operator in xo_functions.cpp
+    *state = *p_game;
     *temp= *state;
-    
+
     if (current_leaf= NULL)//if this conditoin is met, then this is the head node. construct the head node.
     {
             for (int i=0; i<9 i++)
@@ -112,35 +56,35 @@ Tree_Node::Tree_Node(Tree_Node* current_leaf,  game_state* p_game, bool maximize
                     leaves[i]= NULL;
             }
             value = best_value;//assign this nodes value
-        
+
     }
     else if (p_game->victory()!= ' ' || depth> 8 )//base case. This is a terminating node.
     {
         //set all the sub nodes to NULL
         for (int i=0; i<9 i++)
             leaves[i]=NULL;
-        
+
         //set the value of this terminating node depending on who has won
         if (maximize==true);
-           index++; 
+           index++;
         if (p_game->victory()== 'X')
            value= win_value*pow(-1, index);
         if (p_game->victory()== 'O')
            value= win_value*pow(-1, index+1);
-        
-        
+
+
     }
     else //general recursive case. construct this node by calling the next node
     {
          for (int i=0; i<9 ; i++)
           {
-                
+
                 if( temp.make_move(colour, i)==true);//try to make a move. if its an allowed move, then create the next branch of the tree, and store the info about this particular node.
                 {
                     leaves[i]= Tree_Node(this, temp, !maximize, pgoesfirst);//construct the next level down of leaves
                     test_value= leaves[i]->value ;
 
-                    
+
                     //assign a value to this node using minmax  algorithm
                     if (maximize==true)
                         best_value= max( test_value, best_value);
@@ -152,18 +96,18 @@ Tree_Node::Tree_Node(Tree_Node* current_leaf,  game_state* p_game, bool maximize
            }
            value=accumulator;
     }
-}       
+}
 
 //computer fuctions
 void Computer::make_move(Game_State* p_game)
 {
 
     int position=-1;//initially set the move-to-make to a dissallowed value.
-    
+
     double max=0;//play the move that maximizes this value
     game_leaf* trial_leaf= NULL;
-    
-  
+
+
     //loop through searching for moves until a move is found that can be played.
     do
     {
@@ -171,8 +115,8 @@ void Computer::make_move(Game_State* p_game)
         for (int i=0; i<9 ; i++)
         {
             trial_leaf = tree.current_leaf->get_node (tree.curent_leaf->depth+1, i);
-            //if a move in position i is a disslowed move, then trial_leaf will be assigned a value of NULL. 
-                 
+            //if a move in position i is a disslowed move, then trial_leaf will be assigned a value of NULL.
+
             if (trial_leaf!= NULL) //check that its an allowed move
              {
                  if (trial_leaf.get_node_value()> max)//is this move better than all the other moves found so far?
@@ -203,8 +147,8 @@ Computer::Computer(Game_state* p_game, char player_colour, bool playergoesfirst)
         string input;
         getline(cin, input);
         exit (1);
-    
+
     //note that the Tree constructor is allso called (check the prototype declaration to see this call)
-    
+
 }
